Fix package.json. Move code into lib/
