+++ lib/deploy/index.js
