+++ lib/init/index.js
