+++ lib/pack/index.js
