+++ lib/promisify-spawn/index.js
