+++ lib/remove/index.js
