+++ lib/spawn-listener/index.js
