+++ test.js
@@ -2,7 +2,7 @@
 const expect = require('chai').expect;
 
 describe('FaaS provider', () => {
-    const OpenFaaS = require('./faas');
+    const OpenFaaS = require('./lib');
 
     it('creates a new instance', () => {
         const serverless = {
@@ -16,7 +16,7 @@ describe('FaaS provider', () => {
         };
         const opts = {};
         const openFaaS = new OpenFaaS(serverless, opts);
-        expect(faas).to.exist;
-        expect(faas.serverless).to.exist;
+        expect(openFaaS).to.exist;
+        expect(openFaaS.serverless).to.exist;
     });
 });
