+++ tester.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const OpenFaaS = require('./faas');
+const OpenFaaS = require('./lib');
 
 let openFaaS = new OpenFaaS({ pluginManager: { setProvider: () => {}, addPlugin() {} } }, {})
 console.log(openFaaS);
